courseapp/requirements.txt
asgiref==3.8.1
Django==5.1.3
django-ckeditor==6.7.1
django-ckeditor-5==0.2.15
django-debug-toolbar==4.4.6
django-js-asset==2.2.0
djangorestframework==3.15.2
pillow==11.0.0
PyMySQL==1.1.1
sqlparse==0.5.2
tzdata==2024.2
asgiref==3.8.1
Django==5.1.3
drf-yasg==1.21.8
cloudinary==1.41.0
inflection==0.5.1
packaging==24.2
pytz==2024.1
PyYAML==6.0.2
uritemplate==4.1.1
certifi==2024.8.30
django-oauth-toolkit==3.0.1

